Fix conversion warning in qt example

+++ source/examples/qtexample/main.cpp
@@ -216,7 +216,7 @@ public:
 
         ivec2 pos(event->pos().x(), event->pos().y());
 
-        m_nav.scaleAtMouse(pos, -event->delta() * 0.001f);
+        m_nav.scaleAtMouse(pos, static_cast<float>(-event->delta()) * 0.001f);
 
         doneCurrent();
     }
